pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "skiall"
version = "0.1.0"
description = "Cross-device agent skill sync with platform adapters"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
]

[project.scripts]
skiall = "skiall.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/skiall"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
